ARC-1264 Add ServerConfigurationFormSecret component and reference from JenkinsConfigurationForm
